spotify/requirements.txt
discord.py 
spotipy 
pyautogui
python-dotenv